Move code action and completion logic out of the language crate
