+++ crates/multi_buffer/src/multi_buffer.rs
@@ -7,7 +7,6 @@ use collections::{BTreeMap, Bound, HashMap, HashSet};
 use futures::{channel::mpsc, SinkExt};
 use git::diff::DiffHunk;
 use gpui::{AppContext, EventEmitter, Model, ModelContext};
-pub use language::Completion;
 use language::{
     char_kind,
     language_settings::{language_settings, LanguageSettings},
